handle export to ares
